fixed strlen

+++ ex5.c
@@ -273,7 +273,7 @@ int takenSpotsInDB(){
 }
 
 int isStringNumeric(char *s){
-    for (int i = 0; i < strlen(s); i++){
+    for (int i = 0; i < (int)strlen(s); i++){
         if (!(s[i] >= '0' && s[i] <= '9')){
             return 0;
         }
